+++ src/fabscan/scanner/laserscanner/FSHardwareController.py
@@ -6,18 +6,19 @@ __email__ = "info@mariolukas.de"
 
 import logging
 import time
+import threading
 
-from FSLaser import Laser
-from FSLed import Led
 from fabscan.FSConfig import ConfigInterface
 from fabscan.FSSettings import SettingsInterface
-from fabscan.util.FSInject import singleton
+from fabscan.lib.util.FSInject import singleton
 from fabscan.scanner.interfaces.FSHardwareController import FSHardwareControllerInterface
 from fabscan.scanner.interfaces.FSImageProcessor import ImageProcessorInterface
 
-from fabscan.scanner.laserscanner.FSTurntable import Turntable
-from fabscan.scanner.laserscanner.FSCamera import FSCamera
-from fabscan.scanner.laserscanner.FSSerial import FSSerialCom
+from fabscan.scanner.laserscanner.driver.FSTurntable import Turntable
+from fabscan.scanner.laserscanner.driver.FSCamera import FSCamera
+from fabscan.scanner.laserscanner.driver.FSSerial import FSSerialCom
+from fabscan.scanner.laserscanner.driver.FSLaser import Laser
+from fabscan.scanner.laserscanner.driver.FSLed import Led
 
 @singleton(
     config=ConfigInterface,
@@ -52,6 +53,7 @@ class FSHardwareControllerSingleton(FSHardwareControllerInterface):
         self.led.off()
         self.turntable.stop_turning()
         self._logger.debug("Hardware controller initialized...")
+        self.lock = threading.Lock()
 
     def flush(self):
         self.camera.camera_buffer.flush()
@@ -100,13 +102,11 @@ class FSHardwareControllerSingleton(FSHardwareControllerInterface):
         Step size calculated to correspond to num_steps_per_rotation
         Returns resulting image
             '''
-        if color:
-            speed = 800
-        else:
-            speed = 50
 
-        self.turntable.step_interval(steps, speed)
+        self.turntable.step(steps, speed=900)
+        time.sleep(0.7)
         img = self.camera.device.get_frame()
+
         return img