+++ src/fabscan/scanner/laserscanner/driver/FSSerial.py
@@ -9,8 +9,8 @@ import glob
 import serial
 import time
 import logging
-from fabscan.util.FSUtil import FSSystem
-from fabscan.util.FSInject import inject
+from fabscan.lib.util.FSUtil import FSSystem
+from fabscan.lib.util.FSInject import inject
 from fabscan.FSConfig import ConfigInterface
 
 @inject(
@@ -44,12 +44,12 @@ class FSSerialCom():
 
 
     def avr_device_is_available(self):
-        status = FSSystem.run_command("sudo avrdude -p m328p -b "+str(self.flash_baudrate)+" -carduino -P"+str(self._port))
+        status = FSSystem.run_command("sudo avrdude-autoreset -p m328p -b "+str(self.flash_baudrate)+" -carduino -P"+str(self._port))
         return status == 0
 
     def avr_flash(self, fname):
         FSSystem.run_command("wc -l "+str(fname))
-        status = FSSystem.run_command("sudo avrdude -D -V -U flash:w:"+str(fname)+":i -b "+str(self.flash_baudrate)+" -carduino -pm328p -P"+str(self._port))
+        status = FSSystem.run_command("sudo avrdude-autoreset -D -V -U flash:w:"+str(fname)+":i -b "+str(self.flash_baudrate)+" -carduino -pm328p -P"+str(self._port))
         if status != 0:
             self._logger.error("Failed to flash firmware")
         return status == 0
@@ -58,7 +58,7 @@ class FSSerialCom():
         self._logger.debug("Trying to connect Arduino on port: "+str(self._port))
         # open serial port
         try:
-            self._serial = serial.Serial(str(self._port), int(self._baudrate), timeout=1)
+            self._serial = serial.Serial(str(self._port), int(self._baudrate), timeout=0.3)
             time.sleep(1)
         except:
             self._logger.error("Could not open serial port")
@@ -67,7 +67,7 @@ class FSSerialCom():
         self._serial.close()
 
     def _openSerial(self):
-        basedir = os.path.dirname(os.path.dirname(os.path.dirname(os.path.realpath(__file__))))
+        basedir = os.path.dirname(os.path.dirname(os.path.dirname(os.path.dirname(os.path.realpath(__file__)))))
         flash_file_version = max(sorted(glob.iglob(basedir+'/firmware/*.hex'), key=os.path.getctime,reverse=True))
         flash_version_number = os.path.basename(os.path.normpath(os.path.splitext(flash_file_version)[0]))
         self._logger.debug("Latest available firmware version is: "+flash_version_number)
@@ -129,7 +129,8 @@ class FSSerialCom():
                 time.sleep(2) # Wait for FabScan to initialize
                 self._serial.flushInput() # Flush startup text in serial input
                 self.send("M200;")
-                self._serial.readline()
+                self._serial.readline();
+                # receive version number
                 value = self._serial.readline()
                 value = value.strip()
                 if value != "":
@@ -143,20 +144,17 @@ class FSSerialCom():
 
     def send_and_receive(self, message):
         self.send(message)
-        time.sleep(0.1)
+        time.sleep(0.2)
+        response = ""
         while True:
-            try:
-                time.sleep(0.2)
-                command = self._serial.readline()
-                time.sleep(0.2)
-                command = self._serial.readline()
-                self._logger.debug(command.rstrip("\n"))
-                #if state.rstrip("\n") == ">":
-                return command
-            except Exception as e:
-                self._logger.debug(e)
+            response += self._serial.read()
+            if ">" in response:
+                response = response.rstrip(">")
+                response = response.translate(None, '\n\t\r')
+                if response:
+                    self._logger.debug("Command successfully sent: " + response)
                 break
-        time.sleep(0.1)
+
 
     def flush(self):
        self._serial.flushInput()